# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Recordatorios API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "summary": "Register Route",
        "operationId": "register_route_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          },
          "409": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "title": "Type"
                    },
                    "title": {
                      "type": "string",
                      "title": "Title"
                    },
                    "status": {
                      "type": "integer",
                      "title": "Status"
                    },
                    "detail": {
                      "type": "string",
                      "title": "Detail"
                    },
                    "instance": {
                      "type": "string",
                      "title": "Instance"
                    },
                    "traceId": {
                      "type": "string",
                      "title": "Traceid"
                    }
                  },
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem"
                }
              }
            }
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "title": "Type"
                    },
                    "title": {
                      "type": "string",
                      "title": "Title"
                    },
                    "status": {
                      "type": "integer",
                      "title": "Status"
                    },
                    "detail": {
                      "type": "string",
                      "title": "Detail"
                    },
                    "instance": {
                      "type": "string",
                      "title": "Instance"
                    },
                    "traceId": {
                      "type": "string",
                      "title": "Traceid"
                    }
                  },
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem"
                }
              }
            }
          },
          "429": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "title": "Type"
                    },
                    "title": {
                      "type": "string",
                      "title": "Title"
                    },
                    "status": {
                      "type": "integer",
                      "title": "Status"
                    },
                    "detail": {
                      "type": "string",
                      "title": "Detail"
                    },
                    "instance": {
                      "type": "string",
                      "title": "Instance"
                    },
                    "traceId": {
                      "type": "string",
                      "title": "Traceid"
                    }
                  },
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "summary": "Login Route",
        "operationId": "login_route_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "401": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "title": "Type"
                    },
                    "title": {
                      "type": "string",
                      "title": "Title"
                    },
                    "status": {
                      "type": "integer",
                      "title": "Status"
                    },
                    "detail": {
                      "type": "string",
                      "title": "Detail"
                    },
                    "instance": {
                      "type": "string",
                      "title": "Instance"
                    },
                    "traceId": {
                      "type": "string",
                      "title": "Traceid"
                    }
                  },
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem"
                }
              }
            }
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "title": "Type"
                    },
                    "title": {
                      "type": "string",
                      "title": "Title"
                    },
                    "status": {
                      "type": "integer",
                      "title": "Status"
                    },
                    "detail": {
                      "type": "string",
                      "title": "Detail"
                    },
                    "instance": {
                      "type": "string",
                      "title": "Instance"
                    },
                    "traceId": {
                      "type": "string",
                      "title": "Traceid"
                    }
                  },
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem"
                }
              }
            }
          },
          "429": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "title": "Type"
                    },
                    "title": {
                      "type": "string",
                      "title": "Title"
                    },
                    "status": {
                      "type": "integer",
                      "title": "Status"
                    },
                    "detail": {
                      "type": "string",
                      "title": "Detail"
                    },
                    "instance": {
                      "type": "string",
                      "title": "Instance"
                    },
                    "traceId": {
                      "type": "string",
                      "title": "Traceid"
                    }
                  },
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "summary": "Refresh Route",
        "operationId": "refresh_route_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "401": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "title": "Type"
                    },
                    "title": {
                      "type": "string",
                      "title": "Title"
                    },
                    "status": {
                      "type": "integer",
                      "title": "Status"
                    },
                    "detail": {
                      "type": "string",
                      "title": "Detail"
                    },
                    "instance": {
                      "type": "string",
                      "title": "Instance"
                    },
                    "traceId": {
                      "type": "string",
                      "title": "Traceid"
                    }
                  },
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem"
                }
              }
            }
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "title": "Type"
                    },
                    "title": {
                      "type": "string",
                      "title": "Title"
                    },
                    "status": {
                      "type": "integer",
                      "title": "Status"
                    },
                    "detail": {
                      "type": "string",
                      "title": "Detail"
                    },
                    "instance": {
                      "type": "string",
                      "title": "Instance"
                    },
                    "traceId": {
                      "type": "string",
                      "title": "Traceid"
                    }
                  },
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem"
                }
              }
            }
          },
          "429": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "title": "Type"
                    },
                    "title": {
                      "type": "string",
                      "title": "Title"
                    },
                    "status": {
                      "type": "integer",
                      "title": "Status"
                    },
                    "detail": {
                      "type": "string",
                      "title": "Detail"
                    },
                    "instance": {
                      "type": "string",
                      "title": "Instance"
                    },
                    "traceId": {
                      "type": "string",
                      "title": "Traceid"
                    }
                  },
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "summary": "Logout Route",
        "operationId": "logout_route_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "title": "Type"
                    },
                    "title": {
                      "type": "string",
                      "title": "Title"
                    },
                    "status": {
                      "type": "integer",
                      "title": "Status"
                    },
                    "detail": {
                      "type": "string",
                      "title": "Detail"
                    },
                    "instance": {
                      "type": "string",
                      "title": "Instance"
                    },
                    "traceId": {
                      "type": "string",
                      "title": "Traceid"
                    }
                  },
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "summary": "Me Route",
        "operationId": "me_route_api_v1_auth_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          },
          "401": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "traceId": {
                      "title": "Traceid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Credentials": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "Credentials"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "maxLength": 256,
            "minLength": 20,
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest"
      },
...
      "TokenPair": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserPublic"
          },
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "default": 900
          }
        },
        "type": "object",
        "required": [
          "user",
          "access_token",
          "refresh_token"
        ],
        "title": "TokenPair"
      },
...
      "UserPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "email_verified": {
            "type": "boolean",
            "title": "Email Verified"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "email_verified"
        ],
        "title": "UserPublic"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}